Refactor server setup and socket handling
- Moved Express app configuration to a new loader module for better organization and maintainability.
- Simplified the main server file by delegating middleware and route setup to the new express loader.
- Modularized socket event handlers into separate files for improved readability and separation of concerns.
- Implemented user status management and message handling through dedicated utility functions.
- Enhanced error handling and logging across socket operations.
- Updated CORS and security middleware configurations for better security practices.
